+++ apps/web/src/components/Router.tsx
@@ -989,7 +989,7 @@ const createRoute = (locale: Locales, newApplicantDashboard: boolean) =>
                   path: "community-talent",
                   lazy: () =>
                     import(
-                      "../pages/CommunityInterests/CommunityTalent/CommunityTalentPage"
+                      "../pages/CommunityInterests/CommunityTalentPage/CommunityTalentPage"
                     ),
                 },
                 {
