+++ apps/web/src/pages/CommunityInterests/CommunityTalentPage/CommunityTalentTable.tsx
